Move code shared between lite and micro to elementwise_portable.h
